StatisticService.findBy(Topic) separated to StatisticFindService interface
