--- src/main/java/egl/client/service/model/core/StatisticServiceHolder.java
@@ -1,21 +1,0 @@
-package egl.client.service.model.core;
-
-import java.util.Optional;
-
-import egl.client.model.core.profile.Profile;
-import egl.client.model.core.statistic.Result;
-import egl.client.model.core.statistic.TaskStatistic;
-import egl.client.model.core.statistic.TopicStatistic;
-import egl.client.model.core.task.Task;
-import egl.client.model.core.topic.Topic;
-import javafx.beans.property.Property;
-
-public interface StatisticServiceHolder {
-
-    Property<Profile> selectedProfileProperty();
-
-    Optional<TopicStatistic> findBy(Topic localTopic);
-    TaskStatistic findBy(TopicStatistic topicStatistic, Task task);
-
-    void update(TopicStatistic topicStatistic, Task task, Result result);
-}
