+++ src/main/java/egl/client/service/model/local/LocalStatisticService.java
@@ -1,13 +1,13 @@
 package egl.client.service.model.local;
 
-import egl.client.service.model.core.StatisticService;
-import egl.client.service.model.core.StatisticServiceHolder;
+import egl.client.service.model.core.AbstractStatisticService;
+import egl.client.service.model.core.StatisticFindService;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 @Transactional
 @Service
-public class LocalStatisticService extends StatisticService implements StatisticServiceHolder {
+public class LocalStatisticService extends AbstractStatisticService implements StatisticFindService {
 
     public LocalStatisticService(
             LocalProfileService profileService,
